wip ok

+++ blog/database/migrations/2019_10_28_132042_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->timestamps();
             $table->string('nome');
             $table->string('email');
-            $table->integer('celular');
+            $table->bigInteger('celular');
             $table->integer('acesso');
             $table->integer('status')->default(App\Enums\UserStatus::INACTIVE);
         });
